Updated README.md and added HTML, CSS and JavaScript files and images
Updated README.md file with description of the repository and project.
Added index.html, style.css and main.js files. Images for backgrounds,
were also added. Other files and folders include: favicon, a folder
with the background images, a folder with the css styles, and a folder
for the JavaScript.